WIP: rpmautospec plugin
Adds a new plugin to preprocess the RPM specfile prior to building an
SRPM. This plugin will automatically install the `rpmautospec` package
when necessary.

The plugin will first determine if rpmautospec macros are in use. If
not, it will return control with the specfile unmodified. If the
specfile requires processing, it will call out to the
`/usr/bin/rpmautospec process-distgit` command with the `--input
<spec>`, `--output <spec`> and `--git-dir <path>` arguments set
appropriately.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>